+++ src/numbers/ReadonlyRange.ts
@@ -1,3 +1,5 @@
+import Range from './Range';
+
 /**
  * Represents a readonly range between two numbers with a variety of functions
  * that simplify working with number ranges.
@@ -59,7 +61,19 @@ export default interface ReadonlyRange {
      * If this range completely envelops the target range, returns this range's center.
      * If there is no intersection, returns NaN.
      */
-    intersect(range: ReadonlyRange): number;
+    intersectionPoint(range: ReadonlyRange): number;
+    /**
+     * Gets the intersection between this range and `range` as a new Range.
+     * If there is no intersection, returns an empty Range.
+     */
+    intersect(range: ReadonlyRange): Range;
+    /**
+     * Gets the union between this range and `range` as a new Range.
+     * If there is no true union (that is, the two ranges do not intersect),
+     * the returned Range will "bridge the gap", so to speak, acting as if
+     * either of the ranges were large enough to meet the other.
+     */
+    union(range: ReadonlyRange): Range;
     /** Gets the value located at the specified value. For the returned value to be inside this range, the argument should be between 0.0 and 1.0. */
     at(value: number): number;
     /** Gets a relative value between 0.0 and 1.0 to indicate the position of the passed value inside the range. This function is the counter-component to `at()`. */
@@ -70,10 +84,21 @@ export default interface ReadonlyRange {
      * For instance, in a range of -3 to 3, the value 4 would be wrapped to -2.
      */
     wrap(value: number): number;
+    /** 
+    * If this range intersects with `range`, gets the total distance that
+    * this range should be moved so that it no longer intersects with `range`.
+    * 
+    * This function will always choose the shortest distance. The distance
+    * can be negative. If the ranges don't intersect, the return value will
+    * be 0.
+    */
+    getOffset(range: ReadonlyRange): number
+    /** Returns true if the range is empty, that is it spans no distance. */
+    isEmpty(): boolean;
     /** Checks if the given range is identical to this one. */
     equals(range: ReadonlyRange): boolean;
     /** Checks if this range has the given `from` and `to` values. */
     equals(from: number, to: number): boolean;
     /** Clones the range. */
-    clone(): ReadonlyRange;
+    clone(): Range;
 }